Module Ca_certs_nss.Make

Parameters

C : Mirage_clock.PCLOCK

Signature

val authenticator : ?⁠crls:X509.CRL.t list -> ?⁠hash_whitelist:Mirage_crypto.Hash.hash list -> unit -> (X509.Authenticator.t, [> `Msg of string ]) Stdlib.result

authenticator ~crls ~hash_whitelist () is an authenticator with the provided revocation lists, hash_whitelist. The trust anchors are based on the extraction from NSS' certdata.txt.